Erased the tips about the test cases.
Kept the test case under discussion.

Changed some names to include the subsonic word

I discovered an error in the test_PrandtlMeyerExpansion_raises_error_when_deflection_angle_is_over_the_maximum_and_match_is_supersonic test case (old test_pm_raises_error_max_deflection)

Match is important because of it is used to calculate the maximum angle. Even more, if the match is subsonic, then the method raises another diferent exception from the same class.

When changing the match to 0.0, the method raises a diferent exception, but it makes the tests case pass sucesful, and I didn't notice it.

The clue to discover this mistake was the coverage. The code lines that raises the right exception where not executed, but the test passed.

To avoid this mistake again I have added an additional check to verify that the messages of the expected exceptions are valid.